[GHC Upgrade]: Fix Aeson V2.x breaking changes
This commit fixes the breaking changes that come with Aeson V2.x update
while upgrading to GHC 9.2.8 and related package set.

Aeson Changed the internal representation of Object from HashMap -> newer KeyMap.

The choice here was to:-
1. Change the underlying internal representation of `ObjectF`
   from HashMap to KeyMap

2. Do conversions between HashMap and KeyMap
   at endpoints or nodal code that requires it.

In order to keep the changes and downstream breakage minimal,
approach 2 of conversion where required was chosen.